template version and retry_count - i32 to u64 (#374)

+++ cli/golem-cli/src/model/template.rs
@@ -12,7 +12,7 @@ use crate::model::wave::{function_wave_compatible, DisplayNamedFunc, WrapExportF
 #[serde(rename_all = "camelCase")]
 pub struct TemplateView {
     pub template_id: String,
-    pub template_version: i32,
+    pub template_version: u64,
     pub template_name: String,
     pub template_size: i32,
     pub exports: Vec<String>,
